Avoid circular import in network integration
- This prepares the network integration to be usable in http
